requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pytest
pytest-asyncio
respx
google-generativeai
